Revert "fix: 認証情報保持の確認"
This reverts commit 4198f97fe2a8fca51d916292c6b29e1585c903ba.
