--- apps/client/src/contexts/AuthContext.tsx
@@ -1,80 +1,0 @@
-"use client";
-
-import {
-  createContext,
-  useContext,
-  useState,
-  useEffect,
-  ReactNode,
-} from "react";
-import { apiClient, type User } from "@/lib/api";
-
-interface AuthContextType {
-  user: User | null;
-  isLoading: boolean;
-  isAuthenticated: boolean;
-  signIn: () => void;
-  signOut: () => Promise<void>;
-  refetch: () => Promise<void>;
-}
-
-const AuthContext = createContext<AuthContextType | undefined>(undefined);
-
-export function AuthProvider({ children }: { children: ReactNode }) {
-  const [user, setUser] = useState<User | null>(null);
-  const [isLoading, setIsLoading] = useState(true);
-
-  const checkAuth = async () => {
-    try {
-      const response = await apiClient.getCurrentUser();
-      console.log("Auth check response:", response);
-      setUser(response?.user || null);
-    } catch (error) {
-      console.error("Auth check error:", error);
-      setUser(null);
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  useEffect(() => {
-    checkAuth();
-  }, []);
-
-  const signOut = async () => {
-    try {
-      await apiClient.signOut();
-      setUser(null);
-      window.location.href = "/signin";
-    } catch (error) {
-      console.error("Sign out error:", error);
-    }
-  };
-
-  const signIn = () => {
-    window.location.href = `${process.env.NEXT_PUBLIC_API_BASE_URL || "http://localhost:8000"}/api/auth/signin/github`;
-  };
-
-  return (
-    <AuthContext.Provider
-      value={{
-        user,
-        isLoading,
-        isAuthenticated: !!user,
-        signIn,
-        signOut,
-        refetch: checkAuth,
-      }}
-    >
-      {children}
-    </AuthContext.Provider>
-  );
-}
-
-export function useAuth() {
-  const context = useContext(AuthContext);
-  if (context === undefined) {
-    throw new Error("useAuth must be used within an AuthProvider");
-  }
-  return context;
-}
